Migrated entire functionality back in Activity
