+++ app/src/main/java/com/example/insight/activity/mapActivity/MapsActivity.java
@@ -1,58 +1,96 @@
 package com.example.insight.activity.mapActivity;
 
+import androidx.annotation.NonNull;
 import androidx.core.app.ActivityCompat;
 import androidx.fragment.app.FragmentActivity;
 
 import android.content.pm.PackageManager;
-import android.graphics.Camera;
-import android.location.Location;
 import android.os.Bundle;
+import android.util.Log;
 import android.widget.Toast;
 
 import com.example.insight.R;
-import com.example.insight.activity.exception.BusinessException;
-import com.example.insight.service.MapService;
-import com.google.android.gms.location.LocationListener;
-import com.google.android.gms.maps.CameraUpdate;
-import com.google.android.gms.maps.CameraUpdateFactory;
+import com.example.insight.entity.Location;
+import com.example.insight.entity.enums.LocationType;
 import com.google.android.gms.maps.GoogleMap;
 import com.google.android.gms.maps.OnMapReadyCallback;
 import com.google.android.gms.maps.SupportMapFragment;
 import com.google.android.gms.maps.model.LatLng;
+import com.google.android.gms.maps.model.Marker;
 import com.google.android.gms.maps.model.MarkerOptions;
+import com.google.firebase.database.DataSnapshot;
+import com.google.firebase.database.DatabaseError;
+import com.google.firebase.database.DatabaseReference;
+import com.google.firebase.database.FirebaseDatabase;
+import com.google.firebase.database.ValueEventListener;
 
+import java.util.ArrayList;
 import java.util.List;
 
-public class MapsActivity extends FragmentActivity implements OnMapReadyCallback, LocationListener {
+public class MapsActivity extends FragmentActivity implements OnMapReadyCallback {
 
     private GoogleMap mMap;
-    private MapService mapService;
+    private List<Location> locations;
+    private List<Marker> markers;
+    private DatabaseReference database;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
+        this.locations = new ArrayList<>();
+        this.markers = new ArrayList<>();
+        this.database = FirebaseDatabase.getInstance().getReference("locations");
         setContentView(R.layout.activity_maps);
-        // Obtain the SupportMapFragment and get notified when the map is ready to be used.
         SupportMapFragment mapFragment = (SupportMapFragment) getSupportFragmentManager()
                 .findFragmentById(R.id.map);
         mapFragment.getMapAsync(this);
-//        MapFragment mapFragment = (MapFragment) getFragmentManager().findFragmentById(R.id.map);
-//        mapFragment.getMapAsync(this);
     }
 
-    /**
-     * Manipulates the map once available.
-     * This callback is triggered when the map is ready to be used.
-     * This is where we can add markers or lines, add listeners or move the camera. In this case,
-     * we just add a marker near Sydney, Australia.
-     * If Google Play services is not installed on the device, the user will be prompted to install
-     * it inside the SupportMapFragment. This method will only be triggered once the user has
-     * installed Google Play services and returned to the app.
-     */
     @Override
     public void onMapReady(GoogleMap googleMap) {
         this.mMap = googleMap;
-        this.mapService = new MapService(mMap);
+        Log.d("db", "reading");
+        database.orderByChild("id").addValueEventListener(new ValueEventListener() {
+            @Override
+            public void onDataChange(@NonNull DataSnapshot dataSnapshot) {
+                locations = new ArrayList<>();
+                for (DataSnapshot entity : dataSnapshot.getChildren()){
+                    Location toBuild = new Location();
+
+                    toBuild.setId(entity.child("id").getValue().toString());
+                    toBuild.setTitle(entity.child("title").toString());
+                    toBuild.setDescription(entity.child("description").toString());
+                    toBuild.setType(LocationType.valueOf((String)entity.child("type").getValue()));
+
+                    double x = (Double)entity.child("coordinates/latitude").getValue();
+                    double y = (Double)entity.child("coordinates/longitude").getValue();
+
+                    LatLng coordinates = new LatLng(x,y);
+                    toBuild.setCoordinates(coordinates);
+
+                    locations.add(toBuild);
+                }
+
+                mMap.clear();
+
+                for (Location location : locations) {
+                    System.out.println(location.toString());
+                    LatLng coordinates = location.getCoordinates();
+
+                    Marker newMarker = mMap.addMarker(new MarkerOptions()
+                            .position(coordinates)
+                            .title(location.getTitle()));
+                    newMarker.setTag(0);
+                    markers.add(newMarker);
+                }
+            }
+
+            @Override
+            public void onCancelled(@NonNull DatabaseError databaseError) {
+                Log.e("db error", "cancelled", databaseError.toException().getCause());
+            }
+        });
+
         if (mMap == null){
             Toast.makeText(this,"Error loading map!", Toast.LENGTH_LONG).show();
         }
@@ -60,71 +98,5 @@ public class MapsActivity extends FragmentActivity implements OnMapReadyCallback
         if (ActivityCompat.checkSelfPermission(this, android.Manifest.permission.ACCESS_FINE_LOCATION) != PackageManager.PERMISSION_GRANTED && ActivityCompat.checkSelfPermission(this, android.Manifest.permission.ACCESS_COARSE_LOCATION) != PackageManager.PERMISSION_GRANTED) {
             return;
         }
-
-        mMap.clear();
-        mMap = mapService.getMap();
-
-        //mMap = mapService.setMap(mMap);
-
-        CameraUpdate center = CameraUpdateFactory.newLatLng(mapService.getMarkers().get(0).getPosition());
-        CameraUpdate zoom = CameraUpdateFactory.newLatLngZoom(mapService.getMarkers().get(0).getPosition(), 9);
-        mMap.moveCamera(center);
-        mMap.animateCamera(zoom);
-
-//        LatLng clujNapoca = new LatLng(46.770439,23.591423);
-//
-//        CameraUpdate center = CameraUpdateFactory.newLatLng(clujNapoca);
-//        CameraUpdate zoom = CameraUpdateFactory.newLatLngZoom(clujNapoca, 11);
-//
-//        mMap.clear();
-//        mMap = mapService.setMap(mMap);
-//        mMap.moveCamera(center);
-//        mMap.animateCamera(zoom);
-
-//
-//        MarkerOptions mp = new MarkerOptions();
-//        mp.position(clujNapoca);
-//        mp.title("Center");
-//        mMap.addMarker(mp);
-//        mMap.moveCamera(center);
-//        mMap.animateCamera(zoom);
-
-//        mMap.setOnMyLocationChangeListener(new GoogleMap.OnMyLocationChangeListener(){
-//            @Override
-//            public void onMyLocationChange(Location location){
-//                CameraUpdate center = CameraUpdateFactory.newLatLng(new LatLng(location.getLatitude(), location.getLongitude()));
-//                CameraUpdate zoom = CameraUpdateFactory.zoomTo(11);
-//                mMap.clear();
-//
-//                MarkerOptions mp = new MarkerOptions();
-//
-//                mp.position(new LatLng(location.getLatitude(), location.getLongitude()));
-//
-//                mp.title("my position");
-//
-//                mMap.addMarker(mp);
-//                mMap.moveCamera(center);
-//                mMap.animateCamera(zoom);
-//            }
-//        });
-//
-//        // Add a marker in Sydney and move the camera
-//        LatLng sydney = new LatLng(-34, 151);
-//        mMap.addMarker(new MarkerOptions().position(sydney).title("Marker in Sydney"));
-//        mMap.moveCamera(CameraUpdateFactory.newLatLng(sydney));
-
-//        if (googleMap == null){
-//            Toast.makeText(getApplicationContext(),"Sorry! Unable to load map!", Toast.LENGTH_LONG).show();
-//        }
-//
-//        mMap = googleMap;
-//        LatLng clujNapoca = new LatLng(46.770439,23.591423);
-//        mMap.addMarker(new MarkerOptions().position(clujNapoca).title("Suck my Dick"));
-//        mMap.moveCamera(CameraUpdateFactory.newLatLng(clujNapoca));
-    }
-
-    @Override
-    public void onLocationChanged(Location location) {
-
     }
 }
