--- app/src/main/java/com/example/insight/service/MarkerService.java
@@ -1,4 +1,0 @@
-package com.example.insight.service;
-
-public class MarkerService {
-}
